StyleSummary-class {diffobj}	R Documentation

Styling Information for Summaries

Description

Styling Information for Summaries

Slots

container

function applied to entire summary

body

function applied to everything except the actual map portion of the summary

detail

function applied to section showing how many deletions / insertions, etc. occurred

map

function applied to the map portion of the summary

[Package diffobj version 0.3.5 Index]