auto_context {diffobj}	R Documentation

Configure Automatic Context Calculation

Description

Helper functions to help define parameters for selecting an appropriate context value.

Usage

auto_context(
  min = getOption("diffobj.context.auto.min"),
  max = getOption("diffobj.context.auto.max")
)

Arguments

min
integer(1L), positive, set to zero to allow any context

max
integer(1L), set to negative to allow any context

Value

S4 object containing configuration parameters, for use as the context or parameter value in diff* methods

Examples

## `pager="off"` for CRAN compliance; you may omit in normal use
diffChr(letters, letters[-13], context=auto_context(0, 3), pager="off")
diffChr(letters, letters[-13], context=auto_context(0, 10), pager="off")
diffChr(
  letters, letters[-13], context=auto_context(0, 10), line.limit=3L,
  pager="off"
)

[Package diffobj version 0.3.5 Index]