finalizeHtml {diffobj}	R Documentation

Finalizing Methods for HTML Output

Description

Used as the finalizer slot to StyleHtml objects to wrap character output prior to output to device. Used primarily by styles that output to HTML to properly configure HTML page structure, including injecting JS, CSS, etc..

Usage

finalizeHtml(x, ...)

## S4 method for signature 'ANY'
finalizeHtml(x, x.chr, js, ...)

## S4 method for signature 'Diff'
finalizeHtml(x, x.chr, ...)

## S4 method for signature 'DiffSummary'
finalizeHtml(x, x.chr, ...)

Arguments

x
object to finalize

...
arguments to pass on to methods

x.chr
character text representation of x, typically generated with the as.character method for x

js
character javascript code to append to HTML representation

[Package diffobj version 0.3.5 Index]