nchar_html {diffobj}	R Documentation

Count Text Characters in HTML

Description

Very simple implementation that will fail if there are any “>” in the HTML that are not closing tags, and assumes that HTML entities are all one character wide. Also, spaces are counted as one width each because the HTML output is intended to be displayed inside <PRE> tags.

Usage

nchar_html(x, ...)

Arguments

x
character

...
unused for compatibility with internal use

Value

integer(length(x)) with number of characters of each element

Examples

nchar_html("<a href='http:www.domain.com'>hello</a>")

[Package diffobj version 0.3.5 Index]