pager_is_less {diffobj}	R Documentation

Check Whether System Has less as Pager

Description

If getOption(pager) is set to the default value, checks whether Sys.getenv("PAGER") appears to be less by trying to run the pager with the “version” and parsing the output. If getOption(pager) is not the default value, then checks whether it points to the less program by the same mechanism.

Usage

pager_is_less()

Details

Some systems may have less pagers installed that do not respond to the $LESS environment variable. For example, more on at least some versions of OS X is less, but does not actually respond to $LESS. If such as pager is the system pager you will likely end up seeing gibberish in the pager. If this is your use case you will need to set-up a custom pager configuration object that sets the correct system variables (see Pager).

Value

TRUE or FALSE

See Also

Pager

Examples

pager_is_less()

[Package diffobj version 0.3.5 Index]