summary,Diff-method {diffobj}	R Documentation

Summary Method for Diff Objects

Description

Provides high level count of insertions, deletions, and matches, as well as a “map” of where the differences are.

Usage

## S4 method for signature 'Diff'
summary(
  object,
  scale.threshold = 0.1,
  max.lines = 50L,
  width = getOption("width"),
  ...
)

Arguments

object
at Diff object

scale.threshold
numeric(1L) between 0 and 1, how much distortion to allow when creating the summary map, where 0 is none and 1 is as much as needed to fit under max.lines, defaults to 0.1

max.lines
integer(1L) how many lines to allow for the summary map, defaults to 50

width
integer(1L) how many columns wide the output should be, defaults to getOption("width")

...
unused, for compatibility with generic

Details

Sequences of single operations (e.g. "DDDDD") are compressed provided that compressing them does not distort the relative size of the sequence relative to the longest such sequence in the map by more than scale.threshold. Since length 1 sequences cannot be further compressed scale.threshold does not apply to them.

Value

a DiffSummary object ## 'pager="off"' for CRAN compliance; you may omit in normal use summary(diffChr(letters, letters[-c(5, 15)], format="raw", pager="off"))

[Package diffobj version 0.3.5 Index]