[<-,PaletteOfStyles-method {diffobj}	R Documentation

Extract/Replace a Style Class or Object from PaletteOfStyles

Description

Extract/Replace a Style Class or Object from PaletteOfStyles

Usage

## S4 replacement method for signature 'PaletteOfStyles'
x[i, j, ...] <- value

## S4 method for signature 'PaletteOfStyles,ANY,ANY,ANY'
x[i, j, ..., drop = FALSE]

## S4 method for signature 'PaletteOfStyles'
x[[i, j, ..., exact = TRUE]]

Arguments

x
a PaletteOfStyles object

i
numeric, or character corresponding to a valid style format

j
numeric, or character corresponding to a valid style brightness

...
pass a numeric or character corresponding to a valid color.mode

value
a list of Style class or Style objects

drop
TRUE or FALSE, whether to drop dimensions, defaults to FALSE, which is different than generic

exact
passed on to generic

Value

a Style ClassRepresentation object or Style object for [[, and a list of the same for [

See Also

diffPrint for explanations of format, brightness, and color.mode

Examples

pal <- PaletteOfStyles()
pal[["ansi256", "light", "rgb"]]
pal["ansi256", "light", ]
pal["ansi256", "light", "rgb"] <- list(StyleAnsi8NeutralRgb())

[Package diffobj version 0.3.5 Index]