StyleText-class {diffobj}	R Documentation

Character Tokens Used in Diffs

Description

Various character tokens are used throughout diffs to provide visual cues. For example, gutters will contain characters that denote deletions and insertions (< and > by default).

Arguments

gutter.insert
character(1L) text to use as visual cue to indicate whether a diff line is an insertion, defaults to “> ”

gutter.insert.ctd
character(1L) if a diff line is wrapped, the visual cue shifts to this character to indicate wrapping occured

gutter.delete
character(1L) see gutter.insert above

gutter.delete.ctd
character(1L) see gutter.insert.ctd above

gutter.match
character(1L) see gutter.insert above

gutter.match.ctd
character(1L) see gutter.insert.ctd above

gutter.guide
character(1L) see gutter.insert above

gutter.guide.ctd
character(1L) see gutter.insert.ctd above

gutter.fill
character(1L) see gutter.insert above

gutter.fill.ctd
character(1L) see gutter.insert.ctd above

gutter.pad
character(1L) separator between gutter characters and the rest of a line in a diff

pad.col
character(1L) separator between columns in side by side mode

Value

a StyleText S4 object

See Also

Style

[Package diffobj version 0.3.5 Index]