diffObj {diffobj}	R Documentation

Diff Objects

Description

Compare either the printed or str screen representation of R objects depending on which is estimated to produce the most useful diff. The selection process tries to minimize screen lines while maximizing differences shown subject to display constraints. The decision algorithm is likely to evolve over time, so do not rely on this function making a particular selection under specific circumstances. Instead, use diffPrint or diffStr if you require one or the other output.

Usage

diffObj(target, current, ...)

Arguments

target
the reference object

current
the object being compared to target

...
unused, for compatibility of methods with generics

Value

a Diff object; see diffPrint.

See Also

diffPrint for details on the diff* methods, diffStr, diffChr to compare character vectors directly diffDeparse to compare deparsed objects, ses for a minimal and fast diff

Examples

## `pager="off"` for CRAN compliance; you may omit in normal use
diffObj(letters, c(letters[1:10], LETTERS[11:26]), pager="off")
with(mtcars, diffObj(lm(mpg ~ hp)$qr, lm(mpg ~ disp)$qr, pager="off"))

[Package diffobj version 0.3.5 Index]