diffobj_set_def_opts {diffobj}	R Documentation

Set All diffobj Options to Defaults

Description

Used primarily for testing to ensure all options are set to default values.

Usage

diffobj_set_def_opts()

Value

list for use with options that contains values of diffob options before they were forced to defaults

Examples

## Not run: 
  diffobj_set_def_opts()

## End(Not run)

[Package diffobj version 0.3.5 Index]