has_Rdiff {diffobj}	R Documentation

Attempt to Detect Whether diff Utility is Available

Description

Checks whether tools::Rdiff issues a warning when running with useDiff=TRUE and if it does assumes this is because the diff utility is not available. Intended primarily for testing purposes.

Usage

has_Rdiff(test.with = tools::Rdiff)

Arguments

test.with
function to test for diff presence with, typically Rdiff

Value

TRUE or FALSE

Examples

has_Rdiff()

[Package diffobj version 0.3.5 Index]