show,Style-method {diffobj}	R Documentation

Show Method for Style Objects

Description

Display a small sample diff with the Style object styles applied. For ANSI light and dark styles, will also temporarily set the background and foreground colors to ensure they are compatible with the style, even though this is not done in normal output (i.e. if you intend on using a “light” style, you should set your terminal background color to be light or expect sub-optimal rendering).

Usage

## S4 method for signature 'Style'
show(object)

## S4 method for signature 'StyleHtml'
show(object)

Arguments

object
a Style S4 object

Value

NULL, invisibly

Examples

show(StyleAnsi256LightYb())  # assumes ANSI colors supported

[Package diffobj version 0.3.5 Index]