summary,MyersMbaSes-method {diffobj}	R Documentation

Summary Method for Shortest Edit Path

Description

Displays the data required to generate the shortest edit path for comparison between two strings.

Usage

## S4 method for signature 'MyersMbaSes'
summary(object, with.match = FALSE, ...)

Arguments

object
the diff_myers object to display

with.match
logical(1L) whether to show what text the edit command refers to

...
forwarded to the data frame print method used to actually display the data

Value

whatever the data frame print method returns

[Package diffobj version 0.3.5 Index]